[OpenReg] Add Develop Notes for Integrating New Backend into PyTorch
As the title stated.

To facilitate the integration of the new backend, we plan to publish a new
development note that details all the key components, hoping to speed up
the development of other accelerators.

This PR is the beginning of this note, which we will gradually improve and
keep in sync with OpenReg's code.

ghstack-source-id: 2e10e339c3fc257267c5afd70de35ede9c71fc87
Pull-Request-resolved: https://github.com/pytorch/pytorch/pull/158644